adding async retriever methods
